updater-kb/src/requirements.txt
gitPython
jira
atlassian-python-api==3.25.0
PyYAML==6.0